Python/backend-meetmind/requirements.txt
fastapi
uvicorn
pydantic
faster-whisper
python-dotenv
sounddevice
soundfile
numpy
transformers
torch
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
httpx
pytest
openai==0.28